Harrison 2772 - PMCL/PSCH - Support combined maps in ultra initializer. Support obs_date input as integer.
obs_date from L2 maps should be CDF_TIME_TT2000, but isn't yet

+++ imap_l3_processing/ultra/l3/ultra_l3_dependencies.py
@@ -160,7 +160,7 @@ class UltraL3CombinedDependencies:
         dependency_paths = [*u45_pset_paths, *u90_pset_paths, *glows_l3e_pset_paths, u45_map_path,
                             u90_map_path]
         if energy_bin_group_sizes_path:
-            energy_bin_group_sizes = np.loadtxt(energy_bin_group_sizes_path, delimiter=",")
+            energy_bin_group_sizes = np.loadtxt(energy_bin_group_sizes_path, delimiter=",", dtype=np.uint8)
             dependency_paths.append(energy_bin_group_sizes_path)
         else:
             energy_bin_group_sizes = None
